+++ stapi-fastapi/src/stapi_fastapi/models/__init__.py
@@ -1,6 +1,8 @@
-from .opportunity import OpportunityProperties
-from .product import Product, Provider, ProviderRole
-from .shared import Link
+from stapi_pydantic.opportunity import OpportunityProperties
+from stapi_pydantic.product import Provider, ProviderRole
+from stapi_pydantic.shared import Link
+
+from .product import Product
 
 __all__ = [
     "Link",