+++ stapi-fastapi/tests/shared.py
@@ -12,22 +12,19 @@ from geojson_pydantic.types import Position2D
 from httpx import Response
 from pydantic import BaseModel, Field, model_validator
 from pytest import fail
-from stapi_fastapi.models.opportunity import (
+from stapi_fastapi.models.product import Product
+from stapi_pydantic.opportunity import (
     Opportunity,
     OpportunityCollection,
     OpportunityProperties,
     OpportunitySearchRecord,
 )
-from stapi_fastapi.models.order import (
+from stapi_pydantic.order import (
     Order,
     OrderParameters,
     OrderStatus,
 )
-from stapi_fastapi.models.product import (
-    Product,
-    Provider,
-    ProviderRole,
-)
+from stapi_pydantic.product import Provider, ProviderRole
 
 from .backends import (
     mock_create_order,